update SDK to esp_iot_sdk_v1.1.2_15_06_16_p1

+++ hardware/arduino/avr/cores/arduino/WMath.cpp
@@ -27,7 +27,7 @@ extern "C" {
   #include "stdlib.h"
 }
 
-void randomSeed(unsigned int seed)
+void randomSeed(unsigned long seed)
 {
   if (seed != 0) {
     srandom(seed);
